Fixed build errors in the tests
