+++ Source/CollectionViewDataSource.swift
@@ -8,70 +8,6 @@
 
 import UIKit
 
-/// Provides a configuration object for collection view cells.
-public struct CellConfiguration<T> {
-    let cellClass: UICollectionViewCell.Type?
-
-    let cellNib: UINib?
-
-    let presenter: CollectionViewDataSource<T>.CellPresenter?
-
-    /// The cell's reuse identifier
-    public fileprivate(set) var reuseId: String?
-
-    public init(cell: UICollectionViewCell.Type, presenter: CollectionViewDataSource<T>.CellPresenter?) {
-        self.cellClass = cell
-        self.cellNib = nil
-        self.presenter = presenter
-    }
-
-    /// Initializes a cell configuration object.
-    ///
-    /// - Parameters:
-    ///   - reuseId: The cell's reuse identifier
-    ///   - presenter: An optional closure that can be used to inject cell styling and further configuration.
-    public init(cell: UINib, presenter: CollectionViewDataSource<T>.CellPresenter?) {
-        self.cellClass = nil
-        self.cellNib = cell
-        self.presenter = presenter
-    }
-}
-
-/// Provides a configuration object for collection view supplementary views.
-public struct SupplementaryViewConfiguration<T> {
-    let presenter: CollectionViewDataSource<T>.SupplementaryViewPresenter?
-
-    /// The view's reuse id.
-    public fileprivate(set) var reuseId: String?
-
-    /// The kind of supplementary view.
-    public let viewKind: String
-
-    let viewClass: UIView.Type?
-
-    let viewNib: UINib?
-
-    /// Initializes a supplementary view configuration object.
-    ///
-    /// - Parameters:
-    ///   - reuseId: The view's reuse identifier
-    ///   - viewKind: The kind of supplementary view
-    ///   - presenter: An optional closure that can be used to inject view styling and further configuration.
-    public init(view: UIView.Type, viewKind: String, presenter: CollectionViewDataSource<T>.SupplementaryViewPresenter?) {
-        self.presenter = presenter
-        self.viewClass = view
-        self.viewKind = viewKind
-        self.viewNib = nil
-    }
-
-    public init(view: UINib, viewKind: String, presenter: CollectionViewDataSource<T>.SupplementaryViewPresenter?) {
-        self.presenter = presenter
-        self.viewClass = nil
-        self.viewKind = viewKind
-        self.viewNib = view
-    }
-}
-
 public class CollectionViewDataSource<T>: NSObject, UICollectionViewDataSource {
 
     // MARK: Class Types
